Reorganize C++ runtime support
This is done in preparation of future work for additional C++ runtime
support (such as exceptions).
