+++ Kernel/Support/cxa_guard.cpp
@@ -1,5 +1,5 @@
 /**
- * @file cxx.cpp
+ * @file cxa_guard.cpp
  * @author Keeton Feavel (keetonfeavel@cedarville.edu)
  * @brief local static variable guards
  * @version 0.1
@@ -17,36 +17,6 @@
 #include <Panic.hpp>
 #include <stddef.h>
 #include <stdint.h>
-
-/**
- * @brief This function is the global handler for all stack protection. GCC will automatically
- * write the canary code and use this function as the handler for when a smashed stack is detected.
- *
- */
-#if UINTPTR_MAX == UINT32_MAX
-uintptr_t __stack_chk_guard = 0xDEADC0DE;
-#elif UINTPTR_MAX == UINT64_MAX
-uintptr_t __stack_chk_guard = 0xBADBADBADBADBAD1;
-#endif
-
-extern "C"
-{
-
-// Function prototypes (to make compiler happy)
-void __stack_chk_fail(void);
-void __cxa_pure_virtual();
-
-void __stack_chk_fail(void)
-{
-    panic("Smashed stack detected.");
-}
-
-void __cxa_pure_virtual()
-{
-    panic("Pure virtual called!");
-}
-
-}
 
 /**
  * Special thanks to the LLVM maintainers and Apple for their fantastic comments explaining what these
